+++ Tests/Presentation.cs
@@ -72,7 +72,7 @@ namespace OpenXMLOffice.Tests
             PowerPoint powerPoint1 = new("./TestFiles/basic_test.pptx", true);
             Slide Slide = powerPoint1.GetSlideByIndex(0);
             List<Shape> shapes = Slide.FindShapeByText("Slide_1_Shape_1").ToList();
-            shapes[0].ReplaceShape(new Chart(Slide).CreateChart(Global.GlobalConstants.ChartTypes.BAR, CreateDataPayload()));
+            shapes[0].ReplaceShape(new Chart(Slide).CreateChart(Global.GlobalConstants.ColumnChartTypes.CLUSTERED, CreateDataPayload()));
             powerPoint1.SaveAs(string.Format("../../chart-{0}.pptx", DateTime.Now.ToString("yyyy-MM-dd-HH-mm-ss")));
             Assert.IsTrue(true);
         }
@@ -103,18 +103,23 @@ namespace OpenXMLOffice.Tests
             Random random = new();
             DataCell[][] data = new DataCell[5][];
             data[0] = new DataCell[5];
-            for (int col = 0; col < 5; col++)
+            for (int col = 1; col < 5; col++)
             {
                 data[0][col] = new DataCell
                 {
-                    CellValue = $"Heading {col + 1}",
+                    CellValue = $"Series {col}",
                     DataType = CellDataType.STRING
                 };
             }
             for (int row = 1; row < 5; row++)
             {
                 data[row] = new DataCell[5];
-                for (int col = 0; col < 5; col++)
+                data[row][0] = new DataCell
+                {
+                    CellValue = $"Category {row}",
+                    DataType = CellDataType.STRING
+                };
+                for (int col = 1; col < 5; col++)
                 {
                     data[row][col] = new DataCell
                     {
