+++ Tests/Tests.csproj
@@ -2,6 +2,8 @@
 
   <PropertyGroup>
     <RootNamespace>OpenXMLOffice.Tests</RootNamespace>
+    <WeaverCosturaOutputInProjectFolder>true</WeaverCosturaOutputInProjectFolder>
+    <CosturaOutputType>SingleFile</CosturaOutputType>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
@@ -18,7 +20,13 @@
   </ItemGroup>
 
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
-    <Exec Command="xcopy $(ProjectDir)TestFiles $(TargetDir)TestFiles /E /I /Y" />
+    <PropertyGroup Condition="'$(OS)' == 'Windows'">
+      <CopyCommand>xcopy /E /I /Y</CopyCommand>
+    </PropertyGroup>
+    <PropertyGroup Condition="'$(OS)' != 'Windows'">
+      <CopyCommand>cp -r </CopyCommand>
+    </PropertyGroup>
+    <Exec Command="$(CopyCommand) $(ProjectDir)TestFiles $(TargetDir)TestFiles" />
   </Target>
 
 </Project>
